requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.26.1